152 warnings from Warzone in Xcode, 152 warnings in Xcode... Take one down, fix an unsigned/signed comparison warning in mapgrid.cpp, 150 warnings from Warzone in Xcode!
git-svn-id: https://warzone2100.svn.sourceforge.net/svnroot/warzone2100/trunk@9486 4a71c877-e1ca-e34f-864e-861f7616d084

+++ src/mapgrid.cpp
@@ -78,7 +78,7 @@ void gridShutDown(void)
 
 static bool isInRadius(int32_t x, int32_t y, uint32_t radius)
 {
-	return x*x + y*y <= radius*radius;
+	return (uint32_t)(x*x + y*y) <= radius*radius;
 }
 
 // initialise the grid system to start iterating through units that
